requirements.txt
fastapi
uvicorn
python-decouple
urllib3
requests
httpx
outline-vpn-api